ChatBot/requirements.txt
flask-sqlalchemy
flask-login
requests
gunicorn
psycopg2-binary
flask
groq
python-dotenv
